+++ slidy/lib/src/init.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:slidy/src/generate.dart';
+import 'package:slidy/src/package_manager.dart';
 import 'package:slidy/src/utils/utils.dart';
 
 import 'models/app_init_model.dart';
@@ -32,7 +33,9 @@ class Init {
     File(dir.path + "/src/app_module.dart")..createSync(recursive: true)..writeAsStringSync(AppInitModel().appModule(package));
     File(dir.path + "/src/app_bloc.dart")..createSync(recursive: true)..writeAsStringSync(AppInitModel().appBloc());
     File(dir.path + "/src/app_widget.dart")..createSync(recursive: true)..writeAsStringSync(AppInitModel().appWidget(package));
-    
+
+    PackageManager().install(["install", "bloc_pattern", "rxdart", "dio"], false);
+    await Future.delayed(Duration(milliseconds: 800));
     await Generate(["", ""])..module("home/home");
     await Future.delayed(Duration(milliseconds: 800));
     await Generate(["", ""])..component("home", true);